# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Astra VI 3D Cadastre API",
    "description": "SIH26011 spatial workspace and governance prototype. Proposed 3D ULPIN is a prototype extension, not a government standard.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/governance/rules": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Rules",
        "operationId": "rules_api_v1_governance_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/governance/objects/{object_id}": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Case",
        "operationId": "case_api_v1_governance_objects__object_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/governance/right-scopes": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Add Scope",
        "operationId": "add_scope_api_v1_governance_right_scopes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/governance/objects/{object_id}/annotations": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Annotate",
        "operationId": "annotate_api_v1_governance_objects__object_id__annotations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Object Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/governance/review-queue": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Review Queue",
        "operationId": "review_queue_api_v1_governance_review_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/governance/members": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Members",
        "operationId": "members_api_v1_governance_members_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/governance/members/{member_id}": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Update Member",
        "operationId": "update_member_api_v1_governance_members__member_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/governance/objects/{object_id}/history": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Object History",
        "operationId": "object_history_api_v1_governance_objects__object_id__history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Object Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "geometry_versions",
                "workflow",
                "evidence",
                "rights",
                "reviews",
                "annotations",
                "audit",
                "changes"
              ],
              "type": "string",
              "default": "workflow",
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnnotationCreate": {
        "properties": {
          "display_label": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Display Label"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "land_use": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Land Use"
          },
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 5,
            "title": "Reason"
          },
          "expected_previous_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Previous Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_label",
          "evidence_id",
          "reason"
        ],
        "title": "AnnotationCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberUpdate": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "surveyor",
              "officer",
              "viewer"
            ],
            "title": "Role"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "active",
          "reason"
        ],
        "title": "MemberUpdate"
      },
...
      "Payload": {
        "properties": {
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "result"
        ],
        "title": "Payload"
      },
...
      "ScopeCreate": {
        "properties": {
          "right_id": {
            "type": "string",
            "format": "uuid",
            "title": "Right Id"
          },
          "related_object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Related Object Id"
          },
          "reason": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "right_id",
          "related_object_id",
          "reason"
        ],
        "title": "ScopeCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}